organize city elements
