+++ app/src/main/java/ua/sviatkuzbyt/cityweather/ui/screens/forecasts/ForecastScreen.kt
@@ -1,4 +1,4 @@
-package ua.sviatkuzbyt.cityweather.ui.pages.forecasts
+package ua.sviatkuzbyt.cityweather.ui.screens.forecasts
 
 import androidx.compose.animation.Crossfade
 import androidx.compose.foundation.layout.PaddingValues
